split out log from main sockethub.js
