+++ lib/sockethub.js
@@ -1,35 +1,24 @@
 /* global __dirname */
 const nconf      = require('nconf'),
-      path       = require('path'),
       randToken  = require('rand-token'),
       Debug      = require('debug'),
-      express    = require('express'),
-      bodyParser = require('body-parser'),
       kue        = require('kue'),
       dsCrypt    = require('dead-simple-crypt'),
       Store      = require('secure-store-redis'),
       activity   = require('activity-streams')(nconf.get('activity-streams:opts'));
 
-const Worker      = require('./worker.js'),
-      middleware  = require('./middleware.js'),
-      init        = require('./bootstrap/init.js'),
-      Validate    = require('./validate.js'),
-      SR          = require('./shared-resources.js');
+const init            = require('./bootstrap/init.js'),
+      middleware      = require('./middleware.js'),
+      resourceManager = require('./resource-manager.js'),
+      Services        = require('./services.js'),
+      SR              = require('./shared-resources.js'),
+      Validate        = require('./validate.js'),
+      Worker          = require('./worker.js');
 
 const debug        = Debug('sockethub:core  '),
-      rmdebug      = Debug('sockethub:resource-manager'),
       parentId     = randToken.generate(16),
       parentSecret = randToken.generate(64);
 
-let redisCfg = nconf.get('redis');
-if (redisCfg.url) {
-  redisCfg = redisCfg.url;
-}
-
-const queue = kue.createQueue({
-  prefix: 'sockethub:' + parentId + ':queue',
-  redis: redisCfg
-});
 
 function Sockethub() {
   this.counter   = 0;
@@ -45,27 +34,17 @@ Sockethub.prototype.boot = function () {
   if (this.status) {
     return debug('Sockethub.boot() called more than once');
   } else { this.status = true; }
-  debug('redis connection info ', nconf.get('redis'));
 
-  this.app = this.__initExpress();
-  // initialize express and socket.io objects
-  this.http  = require('http').Server(this.app);
-  this.io    = require('socket.io')(this.http, { path: init.path });
+  resourceManager();
 
-  // routes list
-  [
-    'base',
-    'examples'
-  ].map((routeName) => {
-    const route = require( path.join(__dirname, '/../', 'routes', routeName) );
-    return route.setup(this.app);
-  });
+  // start internal and external services
+  services = new Services();
+  this.queue = services.startInternal(parentId);
+  this.io = services.startExternal();
 
-  resourceManagerCyle();
   debug('registering handlers');
   this.__registerJobHandlers();
   this.__registerClientHandlers();
-  this.__startServices();
 };
 
 Sockethub.prototype.__broadcastToSharedPeers = function (origSocket, msg) {
@@ -83,16 +62,6 @@ Sockethub.prototype.__broadcastToSharedPeers = function (origSocket, msg) {
   }
 };
 
-Sockethub.prototype.__initExpress = function () {
-  let app   = express();
-  // templating engines
-  app.set('view engine', 'ejs');
-
-  // use bodyParser
-  app.use(bodyParser.urlencoded({ extended: true }));
-  app.use(bodyParser.json());
-  return app;
-};
 
 // handle job results, from workers, in the queue
 //
@@ -136,8 +105,8 @@ Sockethub.prototype.__processJobResult = function (type) {
 };
 
 Sockethub.prototype.__registerJobHandlers = function () {
-  queue.on('job complete', this.__processJobResult('completed'));
-  queue.on('job failed', this.__processJobResult('failed'));
+  this.queue.on('job complete', this.__processJobResult('completed'));
+  this.queue.on('job failed', this.__processJobResult('failed'));
 };
 
 // handlers from client messages
@@ -205,7 +174,7 @@ Sockethub.prototype.__registerClientHandlers = function () {
 
     socket.on('message', middleware(validate('message'), (msg) => {
       sdebug('queueing incoming job ' + msg.context + ' for socket ' + socket.id);
-      const job = queue.create(socket.id, {
+      const job = this.queue.create(socket.id, {
         title: socket.id + '-' + msg.context + '-' + (msg['@id']) ? msg['@id'] : counter++,
         socket: socket.id,
         msg: dsCrypt.encrypt(JSON.stringify(msg), parentSecret)
@@ -224,74 +193,5 @@ Sockethub.prototype.__registerClientHandlers = function () {
     }));
   });
 };
-
-
-// start up services
-//
-Sockethub.prototype.__startServices = function () {
-  if (nconf.get('kue:enabled')) {
-    // start kue UI
-    kue.app.listen(nconf.get('kue:port'), nconf.get('kue:host'), () => {
-      debug('service queue interface listening on ' + nconf.get('kue:host') + ':'
-        + nconf.get('kue:port'));
-    });
-  }
-
-  // start socket.io
-  this.http.listen(init.port, init.host, () => {
-    debug('sockethub listening on http://' + init.host + ':' + init.port);
-    debug('active platforms: ', [...init.platforms.keys()]);
-  });
-};
-
-
-
-//////
-// resource manager
-//
-function resourceManagerCyle() {
-  rmdebug('initializing resource manager');
-  let reportCount = 0;
-  setInterval(() => {
-    reportCount++;
-    if (reportCount === 4) {
-      rmdebug('sockets: ' + SR.socketConnections.size +
-        ' instances: ' + SR.platformInstances.size);
-      reportCount = 0;
-    }
-
-    for (let platformInstance of SR.platformInstances.values()) {
-      for (let socketId of platformInstance.sockets.values()) {
-        if (!SR.socketConnections.has(socketId)) {
-          rmdebug('removing stale socket reference ' + socketId + ' in platform instance '
-            + platformInstance.id);
-          platformInstance.sockets.delete(socketId);
-        }
-      }
-
-      if (platformInstance.sockets.size <= 0) {
-        if (platformInstance.flaggedForTermination) {
-          // terminate
-          rmdebug(`terminating platform instance ${platformInstance.id} ` +
-            `(flagged for termination: no registered sockets found)`);
-          try {
-            platformInstance.module.cleanup(() => {
-              SR.helpers.removePlatform(platformInstance);
-            });
-          } catch (e) {
-            SR.helpers.removePlatform(platformInstance);
-          }
-        } else {
-          rmdebug(`flagging for termination platform instance ${platformInstance.id} ` +
-            `(no registered sockets found)`);
-          platformInstance.flaggedForTermination = true;
-        }
-      } else {
-        platformInstance.flaggedForTermination = false;
-      }
-    }
-  }, 15000);
-}
-
 
 module.exports = Sockethub;
